fix(pci-object-storage): error boundary
ref: DTCORE-3099
