+++ packages/manager/apps/pci-object-storage/src/pages/Layout.tsx
@@ -1,17 +1,20 @@
-import { Outlet, useRouteError } from 'react-router-dom';
+import { Outlet, useParams } from 'react-router-dom';
 
-import { ShellContext } from '@ovh-ux/manager-react-shell-client';
-import { Suspense, useContext } from 'react';
-import { ErrorBanner } from '@ovh-ux/manager-react-components';
-import { ApiError } from '@ovh-ux/manager-core-api';
-import { useProject } from '@ovh-ux/manager-pci-common';
+import { Suspense } from 'react';
+import { useQuery } from '@tanstack/react-query';
+import { getProjectQuery } from '@ovh-ux/manager-pci-common';
 import ShellRoutingSync from '@/core/ShellRoutingSync';
 import HidePreloader from '@/core/HidePreloader';
 
 import usePageTracking from '@/hooks/usePageTracking';
 
 export default function Layout() {
-  const { isSuccess } = useProject();
+  const { projectId } = useParams();
+  const { isSuccess } = useQuery({
+    ...getProjectQuery(projectId || ''),
+    ...{ retry: false, throwOnError: true },
+  });
+
   usePageTracking();
   return (
     <div className="application">
@@ -27,29 +30,3 @@ export default function Layout() {
     </div>
   );
 }
-
-export const ErrorBoundary = () => {
-  const error = useRouteError() as ApiError;
-  const nav = useContext(ShellContext).shell.navigation;
-
-  const redirectionApplication = 'public-cloud';
-
-  const navigateToHomePage = () => {
-    nav.navigateTo(redirectionApplication, '', {});
-  };
-
-  const reloadPage = () => {
-    nav.reload();
-  };
-  return (
-    <Suspense>
-      <ErrorBanner
-        onReloadPage={reloadPage}
-        onRedirectHome={navigateToHomePage}
-        error={error.response}
-      />
-      <ShellRoutingSync />
-      <HidePreloader />
-    </Suspense>
-  );
-};
